apps/streamlit_app/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
Pillow>=10.1.0
numpy>=1.26.0
streamlit-drawable-canvas>=0.9.3
